Ensure nested dependencies are considered for removal
yarn would only scan the top level packages to generate a list of
candidate packages for removal. When modifying the package tree in
such a way that a nested dependency is no longer required, yarn would
not remove this nested package which could cause subsequent problems
with resolution. The fix was to remove `node_modules` and install
again.

This change will now recursively call `findExtraneousFiles` in
`package-linker`, which means that nested dependencies that are no
longer present in the desired tree are now correctly removed.